+++ ui/components/issuesidebar/styles.go
@@ -1,7 +1,7 @@
 package issuesidebar
 
 import (
-	"github.com/dlvhdr/gh-prs/ui/styles"
+	"github.com/dlvhdr/gh-dash/ui/styles"
 )
 
 var (
