+++ ui/components/issuessection/issuessection.go
@@ -7,14 +7,14 @@ import (
 	"github.com/charmbracelet/bubbles/spinner"
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/charmbracelet/lipgloss"
-	"github.com/dlvhdr/gh-prs/config"
-	"github.com/dlvhdr/gh-prs/data"
-	"github.com/dlvhdr/gh-prs/ui/components/issue"
-	"github.com/dlvhdr/gh-prs/ui/components/section"
-	"github.com/dlvhdr/gh-prs/ui/components/table"
-	"github.com/dlvhdr/gh-prs/ui/constants"
-	"github.com/dlvhdr/gh-prs/ui/context"
-	"github.com/dlvhdr/gh-prs/utils"
+	"github.com/dlvhdr/gh-dash/config"
+	"github.com/dlvhdr/gh-dash/data"
+	"github.com/dlvhdr/gh-dash/ui/components/issue"
+	"github.com/dlvhdr/gh-dash/ui/components/section"
+	"github.com/dlvhdr/gh-dash/ui/components/table"
+	"github.com/dlvhdr/gh-dash/ui/constants"
+	"github.com/dlvhdr/gh-dash/ui/context"
+	"github.com/dlvhdr/gh-dash/utils"
 )
 
 const SectionType = "issues"
