+++ ui/components/listviewport/styles.go
@@ -2,7 +2,7 @@ package listviewport
 
 import (
 	"github.com/charmbracelet/lipgloss"
-	"github.com/dlvhdr/gh-prs/ui/styles"
+	"github.com/dlvhdr/gh-dash/ui/styles"
 )
 
 var (
