+++ ui/ui.go
@@ -6,18 +6,18 @@ import (
 	"github.com/charmbracelet/bubbles/key"
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/charmbracelet/lipgloss"
-	"github.com/dlvhdr/gh-prs/config"
-	"github.com/dlvhdr/gh-prs/data"
-	"github.com/dlvhdr/gh-prs/ui/components/help"
-	"github.com/dlvhdr/gh-prs/ui/components/issuesidebar"
-	"github.com/dlvhdr/gh-prs/ui/components/issuessection"
-	"github.com/dlvhdr/gh-prs/ui/components/prsidebar"
-	"github.com/dlvhdr/gh-prs/ui/components/prssection"
-	"github.com/dlvhdr/gh-prs/ui/components/section"
-	"github.com/dlvhdr/gh-prs/ui/components/sidebar"
-	"github.com/dlvhdr/gh-prs/ui/components/tabs"
-	"github.com/dlvhdr/gh-prs/ui/context"
-	"github.com/dlvhdr/gh-prs/utils"
+	"github.com/dlvhdr/gh-dash/config"
+	"github.com/dlvhdr/gh-dash/data"
+	"github.com/dlvhdr/gh-dash/ui/components/help"
+	"github.com/dlvhdr/gh-dash/ui/components/issuesidebar"
+	"github.com/dlvhdr/gh-dash/ui/components/issuessection"
+	"github.com/dlvhdr/gh-dash/ui/components/prsidebar"
+	"github.com/dlvhdr/gh-dash/ui/components/prssection"
+	"github.com/dlvhdr/gh-dash/ui/components/section"
+	"github.com/dlvhdr/gh-dash/ui/components/sidebar"
+	"github.com/dlvhdr/gh-dash/ui/components/tabs"
+	"github.com/dlvhdr/gh-dash/ui/context"
+	"github.com/dlvhdr/gh-dash/utils"
 )
 
 type Model struct {
